requirements.txt
PyQt6==6.6.1
langchain==0.1.4
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
openai==1.12.0
pylint==3.0.3
bandit==1.7.7
black==24.1.1
pytest==7.4.4
python-magic==0.4.27
tiktoken==0.5.2 